Add files via upload
This project will entail a bakery that sells a list of available goods that the user can select from. Starting off, we will construct an introduction to the user stating what type of business and program we offer. The user will be prompted which bakery items they would like to purchase and the quantity needed. After the quantity is selected, the program will calculate and present the total price to the user.